+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Default threshold level coverage lines in HTML coverage report
 - Hidden table cell showing incompletely covered genes in coverage report
 - Display optional case name on gene coverage report
+- Display error in coverage report when query genes are not found in the database
 - Metrics explanation section on coverage report
 ### Changed
 - Moved helper function from endpoints coverage to crud samples